fix: remove circular dependency
