+++ src/main/resources/application.properties
@@ -4,7 +4,6 @@ jwt.token.secret=${TOKEN_SECRET}
 jwt.token.issuer=${TOKEN_ISSUER}
 logging.level.org.springframework.security=DEBUG
 logging.level.org.springframework.security.web.FilterChainProxy=DEBUG
-spring.main.allow-circular-references=true
 springdoc.swagger-ui.tryItOutEnabled=true
 springdoc.swagger-ui.operationsSorter=alpha
 springdoc.swagger-ui.tagsSorter=alpha
